+++ packages/icons/src/index.ts
@@ -361,3 +361,7 @@ export { ArrowSidebarRightOutline } from './arrow-sidebar-right-outline';
 export { ArrowSidebarLeft } from './arrow-sidebar-left';
 export { ArrowSidebarRight } from './arrow-sidebar-right';
 export { Transformation } from './transformation';
+export { Sharebox } from './sharebox';
+export { NoMembers }  from './no-members';
+export { PositionCenter } from './position-center';
+export { NoTags } from './no-tags';