+++ packages/ui/src/core/drawer/drawer-list.component.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+
+import { intrinsicComponent } from '../../utils/functions';
+import type { DrawerListProps } from './drawer.props';
+import Styled from './drawer.styles';
+
+const DrawerList = intrinsicComponent<DrawerListProps, HTMLUListElement>(
+  ({ children, ...rest }, ref): JSX.Element => (
+    <Styled.List {...rest} ref={ref}>
+      {children}
+    </Styled.List>
+  )
+);
+
+DrawerList.defaultProps = {};
+
+DrawerList.propTypes = {};
+
+export default DrawerList;