+++ packages/ui/src/core/input/input.styles.ts
@@ -108,6 +108,7 @@ const Input = styled.div.attrs({
     box-sizing: border-box;
     column-gap: 6px;
     cursor: text;
+    transition: all 100ms ease-out;
     width: ${fullWidth ? '100%' : '300px'};
     pointer-events: ${disabled ? 'none' : 'auto'};
     background-color: ${getInputBackgroundColor(readOnly, disabled)};
@@ -148,7 +149,7 @@ const Input = styled.div.attrs({
       &:hover {
         background-color: ${isSearchInput
           ? theme.palette[PColor.BackgroundHover]
-          : theme.palette[PColor.TextPrimaryInvert]};
+          : theme.palette[PColor.BackgroundStateless]};
         border: 1px solid ${theme.palette[PColor.BordersPrimaryHover]};
 
         ${Icon} {