+++ packages/ui/stories/core/tooltip.story.tsx
@@ -26,8 +26,7 @@ const defaultArgs = {
   size: Size.Sm,
   arrow: true,
   warning: false,
-  enterDelay: 200,
-  leaveDelay: 200,
+  stayOpen: false,
 };
 
 const BasicTemplate: Story<TooltipProps> = (args) => (