+++ CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.8](https://github.com/scaleflex/ui/compare/2.0.0-beta.7...2.0.0-beta.8) (2022-11-19)
+
+
+### Features
+
+* **Tooltip:** add stayOpen prop ([5e095fd](https://github.com/scaleflex/ui/commit/5e095fd598e61a8cfd847fdc2069da2d5c96fe61))
+
+
+
+
+
+# [2.0.0-beta.7](https://github.com/scaleflex/ui/compare/2.0.0-beta.6...2.0.0-beta.7) (2022-11-17)
+
+
+### Bug Fixes
+
+* update design - T11581 ([1f381b6](https://github.com/scaleflex/ui/commit/1f381b663f3e29298850dbd21aafbd0dc251f3f9))
+
+
+### Features
+
+* **Backdrop:** add Backdrop component ([d441402](https://github.com/scaleflex/ui/commit/d4414028e53badcecf1506d956e6644bd89cdf65))
+* **Divider:** add new divider component ([3b9a5c4](https://github.com/scaleflex/ui/commit/3b9a5c4c65c6d3646bc81ae902ac9515615a1e34))
+* **Drawer:** add Drawer component ([186ac51](https://github.com/scaleflex/ui/commit/186ac51229cbf15e8d485e22aa46ab71a07c3ba6))
+
+
+
+
+
 # [2.0.0-beta.6](https://github.com/scaleflex/ui/compare/2.0.0-beta.5...2.0.0-beta.6) (2022-11-09)